Create internal sealed class CompositeMove : IMove
