+++ Assets/Code/Controller/EnemyInitialization.cs
@@ -1,0 +1,26 @@
+﻿using UnityEngine;
+
+namespace JevLogin
+{
+    internal class EnemyInitialization : IInitialization
+    {
+        private EnemyFactory _enemyFactory;
+        private CompositeMove _enemy;
+
+        public EnemyInitialization(EnemyFactory enemyFactory)
+        {
+            _enemyFactory = enemyFactory;
+            
+        }
+
+        public GameObject GetEnemy()
+        {
+            return _enemyFactory.CreateEnemy();
+        }
+
+        public void Initialization()
+        {
+
+        }
+    }
+}
