+++ Assets/Code/Data/Enemy/EnemyComponents.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using UnityEngine;
+
+namespace JevLogin
+{
+    [Serializable]
+    public sealed class EnemyComponents
+    {
+        public Transform EnemyTransform;
+        public Rigidbody2D Rigidbody2D;
+    }
+}
