--- Assets/Code/Data/EnemyComponents.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using UnityEngine;
-
-namespace JevLogin
-{
-    [Serializable]
-    public sealed class EnemyComponents
-    {
-        public Transform EnemyTransform;
-        public Rigidbody2D Rigidbody2D;
-    }
-}
