--- Assets/Code/Data/EnemyStruct.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using UnityEngine;
-
-namespace JevLogin
-{
-    [Serializable]
-    public struct EnemyStruct
-    {
-        [Range(0, 100)] public float Speed;
-        [Range(0,100)] public int HealthPoint;
-    }
-}
