+++ Assets/Code/Enemy/EnemyModel.cs
@@ -1,0 +1,25 @@
+﻿namespace JevLogin
+{
+    public class EnemyModel
+    {
+        #region Fields
+
+        public EnemyStruct EnemyStruct;
+        public EnemyComponents EnemyComponents;
+        public EnemySettingsData EnemySettingsData;
+
+        #endregion
+
+
+        #region Properties
+
+        public EnemyModel(EnemyStruct enemyStruct, EnemyComponents enemyComponents, EnemySettingsData enemySettings)
+        {
+            EnemyStruct = enemyStruct;
+            EnemyComponents = enemyComponents;
+            EnemySettingsData = enemySettings;
+        }
+
+        #endregion
+    }
+}
