--- Assets/Code/Enemy/EnemyType.cs
@@ -1,9 +1,0 @@
-﻿namespace JevLogin
-{
-    public enum EnemyType
-    {
-        None    = 0,
-        Small   = 1,
-        Big     = 2
-    }
-}
