+++ Assets/Resources/Data/PlayerData.asset
@@ -16,7 +16,7 @@ MonoBehaviour:
     _speed: 5
     _acceleration: 10
     _healthPoint: 100
-    _force: 1000
+    _force: 96
   PlayerComponents:
     SpriteRenderer: {fileID: 0}
     CircleCollider2D: {fileID: 0}
